redundant cast

+++ source/com/bombinggames/wurfelengine/core/map/rendering/RenderBlock.java
@@ -1243,7 +1243,7 @@ public class RenderBlock extends AbstractGameObject {
 	 * Makes every side visible
 	 */
 	public void setUnclipped() {
-		clipping = (byte) 0;
+		clipping = 0;
 	}
 
 	/**
